refactor(table): extract setup logic to useTable per sfc-pattern
